+++ hyperspy/tests/misc/test_hist_tools.py
@@ -23,7 +23,6 @@ import pytest
 
 import hyperspy.api as hs
 from hyperspy.decorators import lazifyTestClass
-from hyperspy.exceptions import VisibleDeprecationWarning
 
 
 def generate_bad_toy_data():
@@ -96,10 +95,3 @@ class TestHistogramBinMethodsBadDataset:
     def test_working_bins(self, bins, size):
         out = self.s1.get_histogram(bins=bins)
         assert out.data.shape == size
-
-    @pytest.mark.parametrize("bins", ["scotts", "freedman"])
-    def test_deprecation_warnings(self, bins):
-        with pytest.warns(
-            VisibleDeprecationWarning, match="has been deprecated and will be removed"
-        ):
-            _ = self.s1.get_histogram(bins=bins)